updated openzeppelin and TutorialToken

+++ contracts/TutorialToken.sol
@@ -1,12 +1,12 @@
 pragma solidity >=0.4.21 <0.7.0;
 
-import "openzeppelin-solidity/contracts/token/ERC20/ERC20.sol";
+import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
 contract TutorialToken is ERC20 {
     string public name = "TutorialToken";
     string public symbol = "TT";
-    uint public decimals = 2;
-    uint public INITIAL_SUPPLY = 12000;
+    uint256 public decimals = 2;
+    uint256 public INITIAL_SUPPLY = 12000;
 
     constructor() public {
         _mint(msg.sender, INITIAL_SUPPLY);
